Initial import from tichy
This add all the minimum files needed to start the launcher and the paroli
applications.

The phone related classes and services have been put into paroli-services
directory and are not part of tichy anymore (this means that it is not possible
anymore to use tichy.phone, tichy.Contact or tichy.Message for example).

The setup file is still missing.