Introduce a state machine mechanism in paroli-contacts and rearrange it.
This is like treat the layers as the states for an application.
The state increases and decreases as the UI screen flows. Basically one
state stands for certain screen, one certain edje obj and layer.
Different screens of an application can use the same layer(state).

By using a layerdict dictionary to keeping track of what layer the
application is at. Then when the app being closed, it knows what edje
objs it's going to be deleted and it won't happen intending to delete an
object that's alread deleted.

I think this is kind state machine mechanism. Though layer value is
magic number. this can be improved. Even we can just use one layer for
one application. That depends on whether the app needs or not.